Fix audience value for authorization
This had been set to the wrong value (not sure where I got it from tbh). Trying to use it gave an "Service not enabled within domain: 616de4f3656d0a021bf12acd" error when trying to exchange tokens. Fixing it, with the value copied from the Quick Start on the Auth0 page. Tested and confirmed it works with cURL.